Fix bug when user closes browser, crashes server.
This might be a bug with websocket itself (however I can't seem to find
a bug report related to it). When a user closes the browser (testing
between firefox and chromium, it seems only chromium/google chrome has
this issue), sometimes a connection request is sent to the server,
instead of a close request. It seems to be more likely to happen after
several send()'s, then closing the browser; as opposed to just opening
the browser and then closing it.'