Drag and drop system now uses Dragula
Switching to dragula allowed me to remove several "hacks" to get the
other drag and drop system working. Dragula just acted like I expected
it to (yay). With the old drag and drop system, I recently started
having issues where it became very very slow (browser hangs) while
dragging. I'm not what caused it. With Dragula, my implementation becomes
slow when there is too many objects on the screen. This is mostly only
noticeable with ff. (like 200+ playlist items and 200+ browser items)
But this is a huge improvement compared to the old drag and drop
implementation.
- Performance improvements
- Bug fixes